Created setting table and changing persistency logic

+++ sqll.h
@@ -21,7 +21,14 @@ inline constexpr const char *CT_VIDEOS =
 
 inline constexpr const char *CT_CHANNELS =
     "CREATE TABLE IF NOT EXISTS Channels"
-    " (Id TEXT PRIMARY KEY, Name TEXT UNIQUE); ";
+    " (Id TEXT PRIMARY KEY, Name TEXT UNIQUE);";
+
+inline constexpr const char *CT_SETTING =
+    "CREATE TABLE IF NOT EXISTS Setting ("
+    "id INTEGER PRIMARY KEY CHECK (id = 1),"
+    "Keep_Feed INTEGER DEFAULT 30,"
+    "Last_updated INTEGER DEFAULT (strftime('%s', 'now'))"
+    ");";
 
 class Sqlite
 {
@@ -169,11 +176,10 @@ public:
             handleError(L"Error Opening Database");
             db = nullptr;
         }
-
+        sqlite3_exec(db, "INSERT OR IGNORE INTO Setting (id, Keep_Feed) VALUES (1, 30);", nullptr, nullptr, nullptr);
         sqlite3_exec(db, "PRAGMA journal_mode=WAL;", nullptr, nullptr, nullptr);
         sqlite3_exec(db, "PRAGMA synchronous=NORMAL;", nullptr, nullptr, nullptr);
         sqlite3_exec(db, "PRAGMA foreign_keys = ON;", nullptr, nullptr, nullptr);
-
     }
 
     ~Sqlite() { close(); }
@@ -193,7 +199,7 @@ public:
         char *err = nullptr;
         if (sqlite3_exec(db, "COMMIT;", nullptr, nullptr, &err) != SQLITE_OK)
         {
-            std::cerr << "COMMIT failed: " << err << "\n";
+            std::wcerr << L"COMMIT failed: " << err << L"\n";
             sqlite3_free(err);
         }
     }
@@ -203,9 +209,35 @@ public:
         char *err = nullptr;
         if (sqlite3_exec(db, "ROLLBACK;", nullptr, nullptr, &err) != SQLITE_OK)
         {
-            std::cerr << "ROLLBACK failed: " << err << "\n";
+            std::wcerr << L"ROLLBACK failed: " << err << L"\n";
             sqlite3_free(err);
         }
+    }
+
+    void saveSettings(int keepFeed = -1)
+    {
+
+        query =
+            "INSERT INTO Setting (id, Keep_Feed, Last_updated) "
+            "VALUES (1, CASE WHEN ?1 = -1 THEN 30 ELSE ?1 END, strftime('%s', 'now')) "
+            "ON CONFLICT(id) DO UPDATE SET "
+            "Keep_Feed = CASE WHEN ?1 = -1 THEN Keep_Feed ELSE ?1 END, "
+            "Last_updated = strftime('%s', 'now');";
+
+        if (sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr) != SQLITE_OK)
+        {
+            std::wcerr << L"Prepare Error: " << (const wchar_t *)sqlite3_errmsg16(db) << std::endl;
+            return;
+        }
+
+        sqlite3_bind_int(stmt, 1, keepFeed);
+
+        int rc = sqlite3_step(stmt);
+
+        if (rc != SQLITE_DONE)
+            handleError(L"Error saveSetting:");
+
+        sqlite3_finalize(stmt);
     }
 
     void createTable(const char *preset)
@@ -213,6 +245,7 @@ public:
         rc = sqlite3_exec(db, preset, nullptr, nullptr, nullptr);
         if (rc != SQLITE_OK)
             handleError(L"Error creating table");
+        return;
     }
 
     int updateChannel(const Channel &ch)
@@ -222,7 +255,7 @@ public:
         rc = sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr);
         if (rc != SQLITE_OK)
         {
-            std::wcerr << L"Errore prepare (update): " << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db)) << std::endl;
+            handleError(L"Errore prepare (updateChannel) ");
             return 0;
         }
 
@@ -232,7 +265,7 @@ public:
         rc = sqlite3_step(stmt);
         if (rc != SQLITE_DONE)
         {
-            std::wcerr << L"Error prepare: " << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db)) << std::endl;
+            handleError(L"Errore updateChannel ");
             return 0;
         }
 
@@ -273,11 +306,12 @@ public:
 
     int removeChannel(const std::wstring &name)
     {
+        beginTransaction();
         query = "DELETE FROM Channels WHERE NAME = ?;";
         rc = sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr);
         if (rc != SQLITE_OK)
         {
-            std::wcerr << L"Error prepare (delete): " << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db)) << std::endl;
+            handleError(L"Error prepare (removeChannel) ");
             return 0;
         }
 
@@ -286,15 +320,18 @@ public:
         rc = sqlite3_step(stmt);
         if (rc != SQLITE_DONE)
         {
-            std::wcerr << L"Error delete: " << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db)) << std::endl;
+            handleError(L"Error removeChannel ");
             return 0;
         }
 
         sqlite3_finalize(stmt);
         if (sqlite3_changes(db) > 0)
         {
+            saveSettings(-1);
+            commitTransaction();
             return 1;
         }
+        commitTransaction();
         return 0;
     }
 
@@ -334,9 +371,7 @@ public:
         rc = sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr);
         if (rc != SQLITE_OK)
         {
-            std::wcerr << L"Error prepare extractChannels: "
-                       << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db))
-                       << std::endl;
+            handleError(L"Error prepare extractChannels:");
             return;
         }
 
@@ -355,79 +390,74 @@ public:
         return;
     }
 
-    void loadVideosAndTrim(std::vector<Video> &vec, const std::wstring &author, uint32_t limit)
+    void loadVideosAndTrim(std::vector<Video> &vec, const std::wstring &author)
     {
+        int limit = genericQuery<int>("SELECT Keep_Feed FROM Setting WHERE id = 1;");
 
         vec.clear();
-        // --- 1) QUERY ---
-        query =
-            "SELECT Timestamp, Title, Author, Id, Short "
-            "FROM Videos "
-            "WHERE Author = ? "
-            "ORDER BY Timestamp DESC;";
 
-        rc = sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr);
-        if (rc != SQLITE_OK)
+        // --- 1) QUERY SELECT ---
+        query = "SELECT Timestamp, Title, Author, Id, Short FROM Videos WHERE Author = ? ORDER BY Timestamp DESC;";
+
+        if (sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr) != SQLITE_OK)
         {
-            std::wcerr << L"Error prepare (loadVideosAndTrim): "
-                       << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db))
-                       << std::endl;
+            handleError(L"Error prepare (loadVideosAndTrim)");
             return;
         }
 
         sqlite3_bind_text16(stmt, 1, author.c_str(), -1, SQLITE_TRANSIENT);
 
-        // --- 2) EXTRACT ---
         while (sqlite3_step(stmt) == SQLITE_ROW)
         {
             Video v{};
-
             v.tp = sqlite3_column_int64(stmt, 0);
-            const void *title = sqlite3_column_text16(stmt, 1);
-            const void *author = sqlite3_column_text16(stmt, 2);
-            const void *id = sqlite3_column_text16(stmt, 3);
-            sqlite3_column_int64(stmt, 4) == 1 ? v.sh = true : v.sh = false;
 
-            // wstring conversion
+            const void *title = sqlite3_column_text16(stmt, 1);
+            const void *auth = sqlite3_column_text16(stmt, 2);
+            const void *vidId = sqlite3_column_text16(stmt, 3);
+
+            v.sh = (sqlite3_column_int(stmt, 4) == 1);
+
             v.title = title ? reinterpret_cast<const wchar_t *>(title) : L"";
-            v.id = id ? reinterpret_cast<const wchar_t *>(id) : L"";
-            v.author = author ? reinterpret_cast<const wchar_t *>(author) : L"";
+            v.author = auth ? reinterpret_cast<const wchar_t *>(auth) : L"";
+            v.id = vidId ? reinterpret_cast<const wchar_t *>(vidId) : L"";
 
             vec.emplace_back(std::move(v));
         }
-
         sqlite3_finalize(stmt);
 
-        // --- 3) DELETE OVER limit VIDEOS ---
-        if (vec.size() > limit)
+        // --- 2) TRIMMING ---
+        if (vec.size() > (size_t)limit)
         {
-            query =
-                "DELETE FROM Videos "
-                "WHERE Author = ? AND Timestamp < ?;";
+              beginTransaction();
 
-            sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr);
-             if (rc != SQLITE_OK)
-        {
-            std::wcerr << L"Error prepare (delete): "
-                       << reinterpret_cast<const wchar_t*>(sqlite3_errmsg16(db))
-                       << std::endl;
-            return;
-        }
+            query = "DELETE FROM Videos WHERE Author = ? AND Timestamp < ?;";
+       
+            rc = sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr);
+            if (rc != SQLITE_OK)
+            {
+                handleError(L"Error prepare (delete videos over limit)");
+                return;
+            }
 
-            // Timestamp limit  (30th) recent video
-            time_t cutoff = vec[limit-1].tp;
+            // Il cutoff last timestamp video
+            long long cutoff = vec[limit - 1].tp;
 
             sqlite3_bind_text16(stmt, 1, author.c_str(), -1, SQLITE_TRANSIENT);
             sqlite3_bind_int64(stmt, 2, cutoff);
 
-            sqlite3_step(stmt);
+            if (sqlite3_step(stmt) != SQLITE_DONE)
+            {
+                handleError(L"Error executing delete trim");
+            }
+
             sqlite3_finalize(stmt);
 
-            // only limit size
+            saveSettings(-1); 
+            commitTransaction();
+
             vec.resize(limit);
         }
-
-        return;
     }
 
     int insertVideosBatch(const std::vector<Video> &videos)
@@ -464,15 +494,18 @@ public:
             }
             else
             {
-                std::wcerr << L"Error batch line: " << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db)) << std::endl;
+                sqlite3_finalize(stmt);
+                handleError(L"Error during batch step");
             }
 
-            // Reset 
+            // Reset
             sqlite3_reset(stmt);
             sqlite3_clear_bindings(stmt);
         }
 
         sqlite3_finalize(stmt);
+        if (counter > 0)
+            saveSettings(-1);
         commitTransaction();
 
         return counter;
@@ -493,7 +526,7 @@ public:
         rc = sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr);
         if (rc != SQLITE_OK)
         {
-            std::wcerr << L"Error prepare (UPSERT): "
+            std::wcerr << L"Error prepare (insertVideo): "
                        << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db))
                        << std::endl;
             return 0;
@@ -509,7 +542,7 @@ public:
         rc = sqlite3_step(stmt);
         if (rc != SQLITE_DONE)
         {
-            std::wcerr << L"Error UPSERT: "
+            std::wcerr << L"Error upsert(insertVideo): "
                        << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db))
                        << std::endl;
             sqlite3_finalize(stmt);
@@ -530,9 +563,7 @@ public:
         rc = sqlite3_prepare_v2(db, sql, -1, &stmt, nullptr);
         if (rc != SQLITE_OK)
         {
-            std::wcerr << L"Error prepare (loadVideosAndTrim): "
-                       << reinterpret_cast<const wchar_t *>(sqlite3_errmsg16(db))
-                       << std::endl;
+            handleError(L"Error prepare (extractVideo24h): ");
             return;
         }
 
@@ -569,7 +600,10 @@ public:
         " LIMIT 1;";
 
         if (sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr) != SQLITE_OK)
+        {
+            handleError(L"Error prepare getVideoBoundaries: ")
             return;
+        }
 
         if (sqlite3_step(stmt) == SQLITE_ROW)
         {
@@ -606,6 +640,7 @@ public:
 
         int totalVideos = genericQuery<int>("SELECT COUNT(*) FROM Videos;");
         int totalChannels = genericQuery<int>("SELECT COUNT(*) FROM Channels;");
+        int keep = genericQuery<int>("SELECT Keep_Feed FROM Setting WHERE id = 1;");
         int last24h = genericQuery<int>("SELECT COUNT(*) FROM Videos WHERE Timestamp > strftime('%s','now') - 86400;");
         int last7d = genericQuery<int>("SELECT COUNT(*) FROM Videos WHERE Timestamp > strftime('%s','now') - 604800;");
         int last2m = genericQuery<int>("SELECT COUNT(*) FROM Videos WHERE Timestamp > strftime('%s','now','-56 days');");
@@ -620,7 +655,7 @@ public:
         int syncr = genericQuery<int>("PRAGMA synchronous;");
         int tables = genericQuery<int>("SELECT COUNT(*) FROM sqlite_master WHERE type='table'");
         int countShort = genericQuery<int>("SELECT COUNT(*) FROM Videos WHERE Short = 1;");
-
+        std::wstring formattedTime = genericQuery<std::wstring>("SELECT strftime('%H:%M %d-%m-%Y', Last_updated, 'unixepoch') FROM Setting WHERE id = 1;");
         std::wstring result = genericQuery<std::wstring>("PRAGMA integrity_check;");
         std::wstring mode = genericQuery<std::wstring>("PRAGMA journal_mode;");
         int fkeys = genericQuery<int>("PRAGMA foreign_keys;");
@@ -650,8 +685,8 @@ public:
             "GROUP BY Author ORDER BY COUNT(*) DESC LIMIT 10;",
             lastWeekC);
 
-        std::wcout << L"\n----------- ACTIVITY LAST 10 DAYS (" << last7d << std::left <<std::setw(47) <<L" Videos) ------------------------- " 
-                    << L"---------------- TOP CHANNELS ------------------- \n\n";
+        std::wcout << L"\n----------- ACTIVITY LAST 10 DAYS (" << last7d << std::left << std::setw(47) << L" Videos) ------------------------- "
+                   << L"---------------- TOP CHANNELS ------------------- \n\n";
         for (int i = 0; i < lastWeek.size(); i++)
         {
             std::wstring bar1(lastWeek[i].second, L'#');
@@ -678,8 +713,8 @@ public:
             "GROUP BY Author ORDER BY COUNT(*) DESC LIMIT 10;",
             lastWeekC);
 
-        std::wcout << L"\n----------- ACTIVITY LAST 2 MONTHS Weekly (" << last2m << std::setw(37) << L" Videos) ---------------- " 
-                    << L"---------------- TOP CHANNELS ------------------- \n\n";
+        std::wcout << L"\n----------- ACTIVITY LAST 2 MONTHS Weekly (" << last2m << std::setw(37) << L" Videos) ---------------- "
+                   << L"---------------- TOP CHANNELS ------------------- \n\n";
         for (int i = 0; i < lastWeek.size(); i++)
         {
             std::wstring bar1((lastWeek[i].second) / 2, L'#');
@@ -696,13 +731,15 @@ public:
 
         std::wcout << L"\n----------- DATABASE STATISTICS ----------------\n\n"
                    << L"Database size: " << size / 1024 << " kb\n"
-                   << L"Schema version " << version << "\n"
+                   << L"Last change: " << formattedTime.c_str() << "\n"
+                   << L"Schema version: " << version << "\n"
                    << L"Integrity: " << result.c_str() << "\n"
                    << L"Freelist pages: " << freelist << "\n"
                    << L"Journal mode: " << mode.c_str() << "\n"
                    << L"Synchronous mode: " << syncr << "\n"
-                   << L"Foreign Keys: " <<fkeys<<"\n"
-                   << L"Total table :" << tables << "\n\n";
+                   << L"Foreign Keys: " << fkeys << "\n"
+                   << L"Limit Videos Saved per Channel: " << keep << "\n"
+                   << L"Total table: " << tables << "\n\n";
     }
 
     void close()
